Adicionando classes de mapeamento de dados e findAll do pedido

+++ src/main/java/com/webservices/ecommerce/dto/request/ItemPedidoRequestDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class ItemPedidoRequestDTO {
 
     private Long produtoId;
-    private Long quantidade;
+    private Integer quantidade;
 }
